Fix: extract_content_text를 commons.views에서 commons.utils로 이동
models가 views를 import하던 레이어 역전을 해소하고 순환 import 위험을 제거한다.
